feat(root): deprecate go api service in place of typescript api service (#132)

+++ solana-programs/programs/drip-v2/src/instructions/init_global_config.rs
@@ -27,7 +27,7 @@ pub struct InitGlobalConfig<'info> {
 #[derive(AnchorSerialize, AnchorDeserialize)]
 pub struct InitGlobalConfigParams {
     pub super_admin: Pubkey,
-    pub default_drip_fee_bps: u64,
+    pub default_drip_fee_bps: u16,
 }
 
 pub fn handle_init_global_config(
